refactor(v1): go modules
