+++ README.md
@@ -1,0 +1,21 @@
+# MiSSy
+
+[![Slack Status](https://microdevs-slackin.herokuapp.com/badge.svg)](https://microdevs-slackin.herokuapp.com) [![Build Status](https://travis-ci.org/microdevs/missy.svg?branch=master)](https://travis-ci.org/microdevs/missy) [![Coverage Status](https://coveralls.io/repos/github/microdevs/missy/badge.svg?branch=master)](https://coveralls.io/github/microdevs/missy?branch=master) [![Go Report Card](https://goreportcard.com/badge/github.com/microdevs/missy)](https://goreportcard.com/report/github.com/microdevs/missy)
+
+MiSSy (Microservice Support System) is an SDK for creating REST services that talk to each other.
+It is build as separate packages (go modules) that can be used separately.
+
+It provides the following functionality:
+
+### Features
+
+* [certificates](/certificates/README.md)
+* [database](/database/README.md)
+* [http](/http/README.md)
+* [log](/log/README.md)
+* [messaging](/messaging/README.md)
+
+
+## How to use it
+
+Update info here...
