+++ certificates/go.mod
@@ -1,0 +1,3 @@
+module github.com/microdevs/missy/certificates
+
+require github.com/pkg/errors v0.8.1
