+++ database/mongo/session.go
@@ -1,12 +1,7 @@
 package mongo
 
 import (
-	"crypto/tls"
-	"net"
 	"time"
-
-	"github.com/globalsign/mgo"
-	"github.com/pkg/errors"
 )
 
 type Config struct {
@@ -17,29 +12,31 @@ type Config struct {
 	SSL      bool          `env:"MONGO_SSL" envDefault:"false"`
 }
 
-func NewSession(c Config) (*mgo.Session, error) {
-	dialInfo := &mgo.DialInfo{
-		Addrs:    []string{c.Host},
-		Timeout:  c.Timeout,
-		Username: c.User,
-		Password: c.Password,
-	}
-	if c.SSL {
-		dialInfo = dialInfoWithSSL(dialInfo)
-	}
-	s, err := mgo.DialWithInfo(dialInfo)
-	if err != nil {
-		return nil, errors.Errorf("couldn't connect to '%s', err: %s", c.Host, err)
-	}
-	return s, nil
-}
+// TODO prepare proper mongo interfaces that will be exposed
 
-func dialInfoWithSSL(dialInfo *mgo.DialInfo) *mgo.DialInfo {
-	tlsConfig := &tls.Config{}
-	tlsConfig.InsecureSkipVerify = true
-	dialInfo.DialServer = func(addr *mgo.ServerAddr) (net.Conn, error) {
-		conn, err := tls.Dial("tcp", addr.String(), tlsConfig)
-		return conn, err
-	}
-	return dialInfo
-}
+//func NewSession(c Config) (*mgo.Session, error) {
+//	dialInfo := &mgo.DialInfo{
+//		Addrs:    []string{c.Host},
+//		Timeout:  c.Timeout,
+//		Username: c.User,
+//		Password: c.Password,
+//	}
+//	if c.SSL {
+//		dialInfo = dialInfoWithSSL(dialInfo)
+//	}
+//	s, err := mgo.DialWithInfo(dialInfo)
+//	if err != nil {
+//		return nil, errors.Errorf("couldn't connect to '%s', err: %s", c.Host, err)
+//	}
+//	return s, nil
+//}
+//
+//func dialInfoWithSSL(dialInfo *mgo.DialInfo) *mgo.DialInfo {
+//	tlsConfig := &tls.Config{}
+//	tlsConfig.InsecureSkipVerify = true
+//	dialInfo.DialServer = func(addr *mgo.ServerAddr) (net.Conn, error) {
+//		conn, err := tls.Dial("tcp", addr.String(), tlsConfig)
+//		return conn, err
+//	}
+//	return dialInfo
+//}
