+++ http/client/client.go
@@ -1,4 +1,4 @@
-package service
+package client
 
 import (
 	"crypto/tls"
