+++ http/go.mod
@@ -1,0 +1,8 @@
+module github.com/microdevs/missy/http
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/go-chi/chi v4.0.0+incompatible
+	github.com/pkg/errors v0.8.1
+	github.com/prometheus/client_golang v0.9.2
+)
