+++ log/go.mod
@@ -1,0 +1,6 @@
+module github.com/microdevs/missy/log
+
+require (
+	github.com/pkg/errors v0.8.1
+	github.com/sirupsen/logrus v1.3.0
+)
