+++ messaging/go.mod
@@ -1,0 +1,8 @@
+module github.com/microdevs/missy/messaging
+
+require (
+	github.com/bouk/monkey v1.0.1
+	github.com/golang/mock v1.2.0
+	github.com/pkg/errors v0.8.1
+	github.com/segmentio/kafka-go v0.2.2
+)
